+++ src/Monitor/Monitor.Test/ScenarioTests/AlertsTests.ps1
@@ -208,3 +208,109 @@ function Test-GetAzureRmAlertHistory
         # No cleanup needed for now
     }
 }
+
+<#
+.SYNOPSIS
+Tests getting the alert rules associated to a resource group.
+#>
+function Test-GetAzureRmMetricAlertRuleV2
+{
+    # Setup
+    $rgname = 'gd-test-ResourceGroup'
+
+    try
+    {
+        $actual = Get-AzMetricAlertRuleV2 -ResourceGroupName $rgname -Name '2112019'
+		Assert-NotNull $actual
+    }
+    finally
+    {
+        # Cleanup
+        # No cleanup needed for now
+    }
+}
+
+<#
+.SYNOPSIS
+Tests creation of Dimension.
+#>
+function Test-NewAzureRmMetricAlertRuleV2DimensionSelection
+{
+
+    try
+    {
+        $actual = New-AzMetricAlertRuleV2DimensionSelection -DimensionName 'Run location' -ValuesToInclude 1,2,3
+		Assert-NotNull $actual
+		Assert-AreEqual $actual.Dimension 'Run location'
+    }
+    finally
+    {
+        # Cleanup
+        # No cleanup needed for now
+    }
+}
+
+<#
+.SYNOPSIS
+Tests creation of Metric Criteria for GenV2 metric alert rule
+#>
+function Test-NewAzureRmMetricAlertRuleV2Criteria
+{
+
+    try
+    {
+        $actual = New-AzMetricAlertRuleV2Criteria -MetricName 'Percentage CPU' -TimeAggregation Total -Operator GreaterThan -Threshold 2
+		Assert-NotNull $actual
+		Assert-AreEqual $actual.MetricName 'Percentage CPU'
+		Assert-AreEqual $actual.Threshold 2
+		Assert-AreEqual $actual.TimeAggregation Total
+    }
+    finally
+    {
+        # Cleanup
+        # No cleanup needed for now
+    }
+}
+
+<#
+.SYNOPSIS
+Tests removing a GenV2 metric alert rule.
+#>
+function Test-RemoveAzureRmAlertRuleV2
+{
+    # Setup
+    $rgname = 'gd-test-ResourceGroup'
+
+    try
+    {
+		Remove-AzMetricAlertRuleV2 -ResourceGroupName $rgname -Name 'temp'
+    }
+    finally
+    {
+        # Cleanup
+        # No cleanup needed for now
+    }
+}
+
+<#
+.SYNOPSIS
+Tests adding a GenV2 metric alert rule.
+#>
+function Test-AddAzureRmMetricAlertRuleV2
+{
+	#Setup
+	$scope = "/subscriptions/80430018-24ee-4b28-a7bd-fb23b5a221d6/resourceGroups/gd-test-ResourceGroup/providers/Microsoft.Compute/virtualMachines/gd-mr-vm1","/subscriptions/80430018-24ee-4b28-a7bd-fb23b5a221d6/resourceGroups/gd-rg-test-2/providers/Microsoft.Compute/virtualMachines/gd-mr-vm2","/subscriptions/80430018-24ee-4b28-a7bd-fb23b5a221d6/resourceGroups/alertstest/providers/Microsoft.Compute/virtualMachines/sr-mr-vm1"
+	$condition = New-AzMetricAlertRuleV2Criteria -MetricName 'Percentage CPU' -TimeAggregation Total -Operator GreaterThan -Threshold 2
+	$actionGroup = New-AzActionGroup -ActionGroupID '/subscriptions/80430018-24ee-4b28-a7bd-fb23b5a221d6/resourcegroups/default-activitylogalerts/providers/microsoft.insights/actiongroups/anashah'
+    try
+    {
+        # Test
+        $actual = Add-AzMetricAlertRuleV2 -Name 'GenV2MetricAlertRule' -ResourceGroupName 'gd-test-ResourceGroup' -WindowSize 00:05:00 -Frequency 00:05:00 -TargetResourceScope $scope -TargetResourceType 'Microsoft.Compute/virtualMachines' -TargetResourceRegion 'southcentralus' -Condition $condition -ActionGroup $actionGroup -Severity 3
+
+    }
+    finally
+    {
+        # Cleanup
+        # No cleanup needed for now
+    }
+}