+++ src/Monitor/Monitor/ChangeLog.md
@@ -18,6 +18,13 @@
         - Additional information about change #1
 -->
 ## Upcoming Release
+  * New cmdlets for GenV2(non classic) metric-based alert rule
+      - New-AzMetricAlertRuleV2DimensionSelection
+      - New-AzMetricAlertRuleV2Criteria
+      - Remove-AzMetricAlertRuleV2
+      - Get-AzMetricAlertRuleV2
+      - Add-AzMetricAlertRuleV2
+  * Updated Monitor SDK to version 0.22.0-preview
 
 ## Version 1.0.0
 * General availability of `Az.Monitor` module